+++ server.js
@@ -1,6 +1,5 @@
 const Koa = require('koa');
 const bodyParser = require('koa-bodyparser');
-const LoginRoutes = require('./routes/Login.routes');
 const UserRoutes = require('./routes/User.routes');
 const AdminCreateUserRoutes = require('./routes/AdminCreateUser.routes');
 const AttendeesPaymentRoutes = require('./routes/AttendeesPayment.routes');
@@ -12,9 +11,6 @@ app.use(cors());
 app.use(bodyParser());
 
 // app.use(LoginRoutes.routes()).use(LoginRoutes.allowedMethods());
-
-app.use(LoginRoutes.routes())
-    .use(LoginRoutes.allowedMethods());
 
 app.use(UserRoutes.routes())
     .use(UserRoutes.allowedMethods());
